factoring out more indexing logic
